requirements.txt
asgiref
blinker
click
colorama
crispy-bootstrap5
crispy-tailwind
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Flask
Flask-Cors
itsdangerous
Jinja2
Markdown
MarkupSafe
mysql-connector-python
pillow
platformdirs
protobuf
psycopg2
psycopg2-binary
sqlparse
typing_extensions
tzdata
virtualenv
Werkzeug